Add bound check for ply from root

+++ src/pvtable.cpp
@@ -20,7 +20,7 @@ inline uint32_t PvTable::m_tableIndex(uint32_t plyFromRoot, uint32_t ply) const
     return plyFromRoot * m_maxPvLength + ply;
 }
 
-void PvTable::updatePv(const Move& move, uint8_t plyFromRoot)
+void PvTable::updatePv(const Move& move, uint32_t plyFromRoot)
 {
     if(plyFromRoot >= m_maxPvLength)
     {
@@ -46,7 +46,7 @@ std::string PvTable::getPvLine()
     return str;
 }
 
-void PvTable::updatePvLength(uint8_t plyFromRoot)
+void PvTable::updatePvLength(uint32_t plyFromRoot)
 {
     if(plyFromRoot >= m_maxPvLength)
     {
@@ -54,5 +54,5 @@ void PvTable::updatePvLength(uint8_t plyFromRoot)
         return;
     }
 
-    m_pvLengths[plyFromRoot] = plyFromRoot;
+    m_pvLengths[plyFromRoot] = static_cast<uint8_t>(plyFromRoot);
 }
