+++ src/pvtable.hpp
@@ -14,8 +14,8 @@ namespace Arcanum
             uint32_t m_tableIndex(uint32_t plyFromRoot, uint32_t ply) const;
         public:
             PvTable(uint32_t maxPvLength);
-            void updatePv(const Move& move, uint8_t plyFromRoot);
-            void updatePvLength(uint8_t plyFromRoot);
+            void updatePv(const Move& move, uint32_t plyFromRoot);
+            void updatePvLength(uint32_t plyFromRoot);
             std::string getPvLine();
     };
 }
